requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
pyserial>=3.5,<4.0
